Add DSL for range aggregation

+++ src/main/java/tech/habegger/elastic/shared/Range.java
@@ -3,16 +3,16 @@ package tech.habegger.elastic.shared;
 import com.fasterxml.jackson.annotation.JsonInclude;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
-public record Range(Integer from, Integer to, String key) {
-    public static Range between(Integer from, Integer to) {
+public record Range(Float from, Float to, String key) {
+    public static Range between(Float from, Float to) {
         return new Range(from, to, null);
     }
 
-    public static Range from(Integer from) {
+    public static Range from(Float from) {
         return new Range(from, null, null);
     }
 
-    public static Range to(Integer to) {
+    public static Range to(Float to) {
         return new Range(null, to, null);
     }
